Minor Code improvements
