+++ ArtistSearch/Common/Enums/InformationType.swift
@@ -1,19 +1,6 @@
 import Foundation
 import UIKit
 
-public enum FilteringType: String {
-    case artist = "artistTerm"
-    
-    static let allFilters: [String]! = ["Artist Name"]
-    
-    public var filteringTypeURL: String {
-        switch self {
-        case .artist:
-            return Constants.baseUrl + Constants.attributeUrl + FilteringType.artist.rawValue + Constants.endUrl
-        }
-    }
-}
-
 public enum InformationType: Int {
     case defaultInformation
     case noResults
