+++ ArtistSearch/Modules/Favorites Search/UI/ViewController/FavoriteSearchViewController.swift
@@ -1,7 +1,7 @@
 import UIKit
 import RealmSwift
 
-class FavoriteSearchViewController: UIViewController {
+class FavoriteSearchViewController: UIViewController, FavoriteSearchViewProtocol {
     @IBOutlet weak var favoriteSearchCollectionView: UICollectionView! {
         didSet {
             favoriteSearchCollectionView.register(FavoriteSearchCollectionViewCell.nib, forCellWithReuseIdentifier: FavoriteSearchCollectionViewCell.reusableIdentifier)
@@ -43,11 +43,6 @@ class FavoriteSearchViewController: UIViewController {
         searchInformationLabel.text = NSLocalizedString("no_favorite_artist_message", comment: "")
         searchInformationImageView.image = #imageLiteral(resourceName: "noFavoriteIcon")
     }
-}
-
-// MARK: - Extensions
-
-extension FavoriteSearchViewController: FavoriteSearchViewProtocol {
     
     func reloadDataInView(with artistData: [ArtistObject]) {
         favoriteArtistData = artistData
@@ -60,89 +55,4 @@ extension FavoriteSearchViewController: FavoriteSearchViewProtocol {
             favoriteSearchCollectionView.isHidden = false
         }
     }
-}
-
-// MARK: - UICollectionViewDelegate methods
-
-extension FavoriteSearchViewController: UICollectionViewDelegate {
-    
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        if !favoriteArtistData.isEmpty, let selectedArtist: ArtistObject = favoriteArtistData?[indexPath.row] {
-            
-            let artist: Artist = Artist.init(artistId: selectedArtist.artistId, 
-                                             primaryGenreName: selectedArtist.primaryGenreName, 
-                                             wrapperType: selectedArtist.wrapperType, 
-                                             artistName: selectedArtist.artistName, 
-                                             artistType: selectedArtist.artistType, 
-                                             artistLinkUrl: selectedArtist.artistLinkUrl, 
-                                             primaryGenreId: selectedArtist.primaryGenreId)
-            presenter?.presentArtistDetail(artist)
-        }
-    }
-}
-
-// MARK: - UICollectionViewDataSource methods
-
-extension FavoriteSearchViewController: UICollectionViewDataSource {
-    
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return 1
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return favoriteArtistData.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell: ArtistCollectionViewCell = collectionView.dequeueReusableCell(withReuseIdentifier: FavoriteSearchCollectionViewCell.reusableIdentifier, 
-                                                                                      for: indexPath) as? ArtistCollectionViewCell else {
-                                                                                        return UICollectionViewCell()
-        }
-        
-        let artistObject: ArtistObject = favoriteArtistData[indexPath.row]
-        
-        let artist: Artist = Artist.init(artistId: artistObject.artistId,
-                                         primaryGenreName: artistObject.primaryGenreName,
-                                         wrapperType: artistObject.wrapperType, 
-                                         artistName: artistObject.artistName, 
-                                         artistType: artistObject.artistType, 
-                                         artistLinkUrl: artistObject.artistLinkUrl, 
-                                         primaryGenreId: artistObject.primaryGenreId)
-        
-        cell.configureCell(with: artist)
-        
-        return cell
-    }
-}
-
-// MARK: - UICollectionViewDelegateFlowLayout methods
-
-extension FavoriteSearchViewController : UICollectionViewDelegateFlowLayout {
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        
-        let width: CGFloat = collectionView.bounds.width / 2 - 4
-        let height: CGFloat = 60.0
-        
-        return CGSize(width: width, height: height)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, 
-                        layout collectionViewLayout: UICollectionViewLayout, 
-                        insetForSectionAt section: Int) -> UIEdgeInsets {
-        return UIEdgeInsets.zero
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, 
-                        layout collectionViewLayout: UICollectionViewLayout, 
-                        minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        return 2
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, 
-                        layout collectionViewLayout: UICollectionViewLayout, 
-                        minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        return 10
-    }
-    
 }
